+++ src/ultimate_rvc/core/generate/song_cover.py
@@ -11,23 +11,20 @@ import logging
 import operator
 import shutil
 from contextlib import suppress
-from functools import cache, reduce
-from itertools import starmap
+from functools import cache
 from pathlib import Path
 from urllib.parse import parse_qs, urlparse
 
 from pydantic import ValidationError
 
-from ultimate_rvc.common import SEPARATOR_MODELS_DIR, VOICE_MODELS_DIR, lazy_import
+from ultimate_rvc.common import SEPARATOR_MODELS_DIR, lazy_import
 from ultimate_rvc.core.common import (
     INTERMEDIATE_AUDIO_BASE_DIR,
     OUTPUT_AUDIO_DIR,
     copy_file_safe,
     display_progress,
     get_file_hash,
-    get_unique_base_path,
     json_dump,
-    json_dumps,
     json_load,
     validate_url,
 )
@@ -38,26 +35,30 @@ from ultimate_rvc.core.exceptions import (
     NotFoundError,
     NotProvidedError,
     UIMessage,
-    VoiceModelNotFoundError,
     YoutubeUrlError,
 )
-from ultimate_rvc.core.typing_extra import (
-    AudioExtInternal,
-    ConvertedVocalsMetaData,
-    DirectoryMetaData,
+from ultimate_rvc.core.generate.common import (
+    convert,
+    get_unique_base_path,
+    mix_audio,
+    validate_all_exist,
+    validate_exists,
+    wavify,
+)
+from ultimate_rvc.core.generate.typing_extra import (
     EffectedVocalsMetaData,
     FileMetaData,
-    MixedSongMetaData,
+    MixedAudioType,
     PitchShiftMetaData,
+    RVCAudioMetaData,
     SeparatedAudioMetaData,
     SourceType,
-    StagedAudioMetaData,
-    WaveifiedAudioMetaData,
 )
 from ultimate_rvc.typing_extra import (
     AudioExt,
     EmbedderModel,
     F0Method,
+    RVCContentType,
     SegmentSize,
     SeparationModel,
     StrPath,
@@ -70,30 +71,24 @@ if TYPE_CHECKING:
 
     import gradio as gr
 
-    import ffmpeg
     import pedalboard
     import pedalboard._pedalboard as pedalboard_
     import pedalboard.io as pedalboard_io
-    import pydub
     import soundfile as sf
     import sox
     import static_ffmpeg
     import static_sox
 
-    # NOTE the only reason these are imported here is so we can annotate
-    # the return value of the two functions below
+    # NOTE the only reason this is imported here is so we can annotate
+    # the return type of the _get_audio_separator function
     from audio_separator.separator import Separator
-
-    from ultimate_rvc.rvc.infer.infer import VoiceConverter
 
 else:
     yt_dlp = lazy_import("yt_dlp")
 
-    ffmpeg = lazy_import("ffmpeg")
     pedalboard = lazy_import("pedalboard")
     pedalboard_ = lazy_import("pedalboard._pedalboard")
     pedalboard_io = lazy_import("pedalboard.io")
-    pydub = lazy_import("pydub")
     sf = lazy_import("soundfile")
     sox = lazy_import("sox")
     static_ffmpeg = lazy_import("static_ffmpeg")
@@ -146,23 +141,6 @@ def _get_audio_separator(
             "enable_denoise": False,
         },
     )
-
-
-@cache
-def _get_voice_converter() -> VoiceConverter:
-    """
-    Get a voice converter.
-
-    Returns
-    -------
-    VoiceConverter
-        A voice converter.
-
-    """
-    static_sox.add_paths()
-    from ultimate_rvc.rvc.infer.infer import VoiceConverter  # noqa: PLC0415
-
-    return VoiceConverter()
 
 
 def initialize_audio_separator(progress_bar: gr.Progress | None = None) -> None:
@@ -319,7 +297,7 @@ def _get_model_name(
         return model_name
     converted_vocals_dict = json_load(converted_vocals_json_path)
     try:
-        converted_vocals_metadata = ConvertedVocalsMetaData.model_validate(
+        converted_vocals_metadata = RVCAudioMetaData.model_validate(
             converted_vocals_dict,
         )
     except ValidationError:
@@ -602,94 +580,6 @@ def retrieve_song(
     return song_path, song_dir_path
 
 
-def _validate_exists(
-    identifier: StrPath | None,
-    entity: Entity,
-) -> Path:
-    """
-    Validate that the provided identifier is not none and that it
-    identifies an existing entity, which can be either a voice model,
-    a song directory or an audio track.
-
-    Parameters
-    ----------
-    identifier : StrPath | None
-        The identifier to validate.
-    entity : Entity
-        The entity that the identifier should identify.
-
-    Returns
-    -------
-    Path
-        The path to the identified entity.
-
-    Raises
-    ------
-    NotProvidedError
-        If the identifier is None.
-    NotFoundError
-        If the identifier does not identify an existing entity.
-    VoiceModelNotFoundError
-        If the identifier does not identify an existing voice model.
-    NotImplementedError
-        If the provided entity is not supported.
-
-    """
-    match entity:
-        case Entity.MODEL_NAME:
-            if not identifier:
-                raise NotProvidedError(entity=entity, ui_msg=UIMessage.NO_VOICE_MODEL)
-            path = VOICE_MODELS_DIR / identifier
-            if not path.is_dir():
-                raise VoiceModelNotFoundError(str(identifier))
-        case Entity.SONG_DIR | Entity.EMBEDDER_MODEL_CUSTOM:
-            ui_msg = UIMessage.NO_SONG_DIR if entity == Entity.SONG_DIR else None
-            if not identifier:
-                raise NotProvidedError(entity=entity, ui_msg=ui_msg)
-            path = Path(identifier)
-            if not path.is_dir():
-                raise NotFoundError(entity=entity, location=path)
-        case (
-            Entity.SONG
-            | Entity.AUDIO_TRACK
-            | Entity.VOCALS_TRACK
-            | Entity.INSTRUMENTALS_TRACK
-            | Entity.MAIN_VOCALS_TRACK
-            | Entity.BACKUP_VOCALS_TRACK
-        ):
-            if not identifier:
-                raise NotProvidedError(entity=entity)
-            path = Path(identifier)
-            if not path.is_file():
-                raise NotFoundError(entity=entity, location=path)
-        case _:
-            error_msg = f"Entity {entity} not supported."
-            raise NotImplementedError(error_msg)
-    return path
-
-
-def _validate_all_exist(
-    identifier_entity_pairs: Sequence[tuple[StrPath | None, Entity]],
-) -> list[Path]:
-    """
-    Validate that all provided identifiers are not none and that they
-    identify existing entities, which can be either voice models, song
-    directories or audio tracks.
-
-    Parameters
-    ----------
-    identifier_entity_pairs : Sequence[tuple[StrPath, Entity]]
-        The pairs of identifiers and entities to validate.
-
-    Returns
-    -------
-    list[Path]
-        The paths to the identified entities.
-
-    """
-    return list(starmap(_validate_exists, identifier_entity_pairs))
-
-
 def separate_audio(
     audio_track: StrPath,
     song_dir: StrPath,
@@ -728,7 +618,7 @@ def separate_audio(
         The path to the separated secondary stem.
 
     """
-    audio_path, song_dir_path = _validate_all_exist(
+    audio_path, song_dir_path = validate_all_exist(
         [(audio_track, Entity.AUDIO_TRACK), (song_dir, Entity.SONG_DIR)],
     )
 
@@ -774,335 +664,6 @@ def separate_audio(
         json_dump(args_dict, secondary_json_path)
 
     return primary_path, secondary_path
-
-
-def _get_rvc_files(model_name: str) -> tuple[Path, Path | None]:
-    """
-    Get the RVC model file and potential index file of a voice model.
-
-    Parameters
-    ----------
-    model_name : str
-        The name of the voice model to get the RVC files of.
-
-    Returns
-    -------
-    model_file : Path
-        The path to the RVC model file.
-    index_file : Path | None
-        The path to the RVC index file, if it exists.
-
-    Raises
-    ------
-    NotFoundError
-        If no model file exists in the voice model directory.
-
-
-    """
-    model_dir_path = _validate_exists(model_name, Entity.MODEL_NAME)
-    file_path_map = {
-        ext: path
-        for path in model_dir_path.iterdir()
-        for ext in [".pth", ".index"]
-        if ext == path.suffix
-    }
-
-    if ".pth" not in file_path_map:
-        raise NotFoundError(
-            entity=Entity.MODEL_FILE,
-            location=model_dir_path,
-            is_path=False,
-        )
-
-    model_file = file_path_map[".pth"]
-    index_file = file_path_map.get(".index")
-    return model_file, index_file
-
-
-def to_wav(
-    audio_track: StrPath,
-    song_dir: StrPath,
-    prefix: str,
-    accepted_formats: set[AudioExt] | None = None,
-    progress_bar: gr.Progress | None = None,
-    percentage: float = 0.5,
-) -> Path:
-    """
-    Convert a given audio track to wav format if its current format is
-    one of the given accepted formats.
-
-    Parameters
-    ----------
-    audio_track : StrPath
-        The path to the audio track to convert.
-    song_dir : StrPath
-        The path to the song directory where the converted audio track
-        will be saved.
-    prefix : str
-        The prefix to use for the name of the converted audio track.
-    accepted_formats : set[AudioExt], optional
-        The audio formats to accept for conversion. If None, the
-        accepted formats are mp3, ogg, flac, m4a and aac.
-    progress_bar : gr.Progress, optional
-        Gradio progress bar to update.
-    percentage : float, default=0.5
-        Percentage to display in the progress bar.
-
-    Returns
-    -------
-    Path
-        The path to the audio track in wav format or the original audio
-        track if it is not in one of the accepted formats.
-
-    """
-    static_ffmpeg.add_paths()
-    # NOTE have to import pydub.utils here because
-    # 1. when using gradio pydub.utils is not available if only pydub
-    #    itself is imported.
-    # 2. when using cli static_ffmpeg must be imported before
-    #    pydub.utils is imported.
-    import pydub.utils as pydub_utils  # noqa: PLC0415
-
-    if accepted_formats is None:
-        accepted_formats = set(AudioExt) - {AudioExt.WAV}
-
-    audio_path, song_dir_path = _validate_all_exist(
-        [(audio_track, Entity.AUDIO_TRACK), (song_dir, Entity.SONG_DIR)],
-    )
-
-    wav_path = audio_path
-
-    song_info = pydub_utils.mediainfo(str(audio_path))
-    logger.info("Song Info:\n%s", json_dumps(song_info))
-    if any(
-        (
-            accepted_format in song_info["format_name"]
-            if accepted_format == AudioExt.M4A
-            else accepted_format == song_info["format_name"]
-        )
-        for accepted_format in accepted_formats
-    ):
-        args_dict = WaveifiedAudioMetaData(
-            audio_track=FileMetaData(
-                name=audio_path.name,
-                hash_id=get_file_hash(audio_path),
-            ),
-        ).model_dump()
-
-        paths = [
-            get_unique_base_path(
-                song_dir_path,
-                prefix,
-                args_dict,
-            ).with_suffix(suffix)
-            for suffix in [".wav", ".json"]
-        ]
-        wav_path, wav_json_path = paths
-        if not all(path.exists() for path in paths):
-            display_progress(
-                "[~] Converting audio track to wav format...",
-                percentage,
-                progress_bar,
-            )
-
-            _, stderr = (
-                ffmpeg.input(audio_path)
-                .output(filename=wav_path, f="wav")
-                .run(
-                    overwrite_output=True,
-                    quiet=True,
-                )
-            )
-            logger.info("FFmpeg stderr:\n%s", stderr.decode("utf-8"))
-            json_dump(args_dict, wav_json_path)
-
-    return wav_path
-
-
-def convert(
-    vocals_track: StrPath,
-    song_dir: StrPath,
-    model_name: str,
-    n_octaves: int = 0,
-    n_semitones: int = 0,
-    f0_methods: Sequence[F0Method] | None = None,
-    index_rate: float = 0.5,
-    filter_radius: int = 3,
-    rms_mix_rate: float = 0.25,
-    protect_rate: float = 0.33,
-    hop_length: int = 128,
-    split_vocals: bool = False,
-    autotune_vocals: bool = False,
-    autotune_strength: float = 1.0,
-    clean_vocals: bool = False,
-    clean_strength: float = 0.7,
-    embedder_model: EmbedderModel = EmbedderModel.CONTENTVEC,
-    embedder_model_custom: StrPath | None = None,
-    sid: int = 0,
-    progress_bar: gr.Progress | None = None,
-    percentage: float = 0.5,
-) -> Path:
-    """
-    Convert a vocals track using a voice model.
-
-    Parameters
-    ----------
-    vocals_track : StrPath
-        The path to the vocals track to convert.
-    song_dir : StrPath
-        The path to the song directory where the converted vocals track
-        will be saved.
-    model_name : str
-        The name of the model to use for vocal conversion.
-    n_octaves : int, default=0
-        The number of octaves to pitch-shift the converted vocals by.
-    n_semitones : int, default=0
-        The number of semitones to pitch-shift the converted vocals by.
-    f0_methods : Sequence[F0Method], optional
-        The methods to use for pitch extraction. If None, the method
-        used is rmvpe.
-    index_rate : float, default=0.5
-        The influence of the index file on the vocal conversion.
-    filter_radius : int, default=3
-        The filter radius to use for the vocal conversion.
-    rms_mix_rate : float, default=0.25
-        The blending rate of the volume envelope of the converted
-        vocals.
-    protect_rate : float, default=0.33
-        The protection rate for consonants and breathing sounds.
-    hop_length : int, default=128
-        The hop length to use for CREPE-based pitch extraction.
-    split_vocals : bool, default=False
-        Whether to split the vocals track into smaller segments before
-        converting it.
-    autotune_vocals : bool, default=False
-        Whether to apply autotune to the converted vocals.
-    autotune_strength : float, default=1.0
-        The strength of the autotune to apply to the converted vocals.
-    clean_vocals : bool, default=False
-        Whether to clean the converted vocals.
-    clean_strength : float, default=0.7
-        The intensity of the cleaning to apply to the converted vocals.
-    embedder_model : EmbedderModel, default=EmbedderModel.CONTENTVEC
-        The model to use for generating speaker embeddings.
-    embedder_model_custom : StrPath, optional
-        The path to a directory with a custom model to use for
-        generating speaker embeddings.
-    sid : int, default=0
-        The speaker id to use for multi-speaker models.
-    progress_bar : gr.Progress, optional
-        Gradio progress bar to update.
-    percentage : float, default=0.5
-        Percentage to display in the progress bar.
-
-    Returns
-    -------
-    Path
-        The path to the converted vocals track.
-
-    """
-    vocals_path, song_dir_path, _ = _validate_all_exist(
-        [
-            (vocals_track, Entity.VOCALS_TRACK),
-            (song_dir, Entity.SONG_DIR),
-            (model_name, Entity.MODEL_NAME),
-        ],
-    )
-    if embedder_model == EmbedderModel.CUSTOM:
-        embedder_model_custom = _validate_exists(
-            embedder_model_custom,
-            Entity.EMBEDDER_MODEL_CUSTOM,
-        )
-
-    vocals_path = to_wav(
-        vocals_path,
-        song_dir_path,
-        "20_Input",
-        accepted_formats={AudioExt.M4A, AudioExt.AAC},
-        progress_bar=progress_bar,
-        percentage=percentage,
-    )
-
-    n_semitones = n_octaves * 12 + n_semitones
-    f0_methods_set = set(f0_methods) if f0_methods else {F0Method.RMVPE}
-
-    args_dict = ConvertedVocalsMetaData(
-        vocals_track=FileMetaData(
-            name=vocals_path.name,
-            hash_id=get_file_hash(vocals_path),
-        ),
-        model_name=model_name,
-        n_semitones=n_semitones,
-        f0_methods=sorted(f0_methods_set),
-        index_rate=index_rate,
-        filter_radius=filter_radius,
-        rms_mix_rate=rms_mix_rate,
-        protect_rate=protect_rate,
-        hop_length=hop_length,
-        split_vocals=split_vocals,
-        autotune_vocals=autotune_vocals,
-        autotune_strength=autotune_strength,
-        clean_vocals=clean_vocals,
-        clean_strength=clean_strength,
-        embedder_model=embedder_model,
-        embedder_model_custom=(
-            DirectoryMetaData(
-                name=Path(embedder_model_custom).name,
-                path=str(embedder_model_custom),
-            )
-            if embedder_model_custom is not None
-            else None
-        ),
-        sid=sid,
-    ).model_dump()
-
-    paths = [
-        get_unique_base_path(
-            song_dir_path,
-            "21_Vocals_Converted",
-            args_dict,
-        ).with_suffix(suffix)
-        for suffix in [".wav", ".json"]
-    ]
-
-    converted_vocals_path, converted_vocals_json_path = paths
-
-    if not all(path.exists() for path in paths):
-        display_progress("[~] Converting vocals using RVC...", percentage, progress_bar)
-
-        rvc_model_path, rvc_index_path = _get_rvc_files(model_name)
-
-        voice_converter = _get_voice_converter()
-
-        voice_converter.convert_audio(
-            audio_input_path=str(vocals_path),
-            audio_output_path=str(converted_vocals_path),
-            model_path=str(rvc_model_path),
-            index_path=str(rvc_index_path) if rvc_index_path else "",
-            pitch=n_semitones,
-            f0_methods=f0_methods_set,
-            index_rate=index_rate,
-            volume_envelope=rms_mix_rate,
-            protect=protect_rate,
-            hop_length=hop_length,
-            split_audio=split_vocals,
-            f0_autotune=autotune_vocals,
-            f0_autotune_strength=autotune_strength,
-            filter_radius=filter_radius,
-            embedder_model=embedder_model,
-            embedder_model_custom=(
-                str(embedder_model_custom)
-                if embedder_model_custom is not None
-                else None
-            ),
-            clean_audio=clean_vocals,
-            clean_strength=clean_strength,
-            post_process=False,
-            resample_sr=0,
-            sid=sid,
-        )
-        json_dump(args_dict, converted_vocals_json_path)
-    return converted_vocals_path
 
 
 def _add_effects(
@@ -1202,11 +763,11 @@ def postprocess(
         The path to the effected vocals track.
 
     """
-    vocals_path, song_dir_path = _validate_all_exist(
+    vocals_path, song_dir_path = validate_all_exist(
         [(vocals_track, Entity.VOCALS_TRACK), (song_dir, Entity.SONG_DIR)],
     )
 
-    vocals_path = to_wav(
+    vocals_path = wavify(
         vocals_path,
         song_dir_path,
         "30_Input",
@@ -1310,11 +871,11 @@ def pitch_shift(
         The path to the pitch-shifted audio track.
 
     """
-    audio_path, song_dir_path = _validate_all_exist(
+    audio_path, song_dir_path = validate_all_exist(
         [(audio_track, Entity.AUDIO_TRACK), (song_dir, Entity.SONG_DIR)],
     )
 
-    audio_path = to_wav(
+    audio_path = wavify(
         audio_path,
         song_dir_path,
         "40_Input",
@@ -1353,74 +914,12 @@ def pitch_shift(
     return shifted_audio_path
 
 
-def _to_internal(audio_ext: AudioExt) -> AudioExtInternal:
-    """
-    Map an audio extension to an internally recognized format.
-
-    Parameters
-    ----------
-    audio_ext : AudioExt
-        The audio extension to map.
-
-    Returns
-    -------
-    AudioExtInternal
-        The internal audio extension.
-
-    """
-    match audio_ext:
-        case AudioExt.M4A:
-            return AudioExtInternal.IPOD
-        case AudioExt.AAC:
-            return AudioExtInternal.ADTS
-        case _:
-            return AudioExtInternal(audio_ext)
-
-
-def _mix_song(
-    audio_track_gain_pairs: Sequence[tuple[StrPath, int]],
-    output_file: StrPath,
-    output_sr: int = 44100,
-    output_format: AudioExt = AudioExt.MP3,
-) -> None:
-    """
-    Mix multiple audio tracks to create a song.
-
-    Parameters
-    ----------
-    audio_track_gain_pairs : Sequence[tuple[StrPath, int]]
-        A sequence of pairs each containing the path to an audio track
-        and the gain to apply to it.
-    output_file : StrPath
-        The path to the file to save the mixed song to.
-    output_sr : int, default=44100
-        The sample rate of the mixed song.
-    output_format : AudioExt, default=AudioExt.MP3
-        The audio format of the mixed song.
-
-    """
-    static_ffmpeg.add_paths()
-    mixed_audio = reduce(
-        lambda a1, a2: a1.overlay(a2),
-        [
-            pydub.AudioSegment.from_wav(audio_track) + gain
-            for audio_track, gain in audio_track_gain_pairs
-        ],
-    )
-    mixed_audio_resampled = mixed_audio.set_frame_rate(output_sr)
-    mixed_audio_resampled.export(
-        output_file,
-        format=_to_internal(output_format),
-    )
-
-
 def mix_song(
     audio_track_gain_pairs: Sequence[tuple[StrPath, int]],
     song_dir: StrPath,
     output_sr: int = 44100,
     output_format: AudioExt = AudioExt.MP3,
     output_name: str | None = None,
-    display_msg: str = "[~] Mixing audio tracks...",
     progress_bar: gr.Progress | None = None,
     percentage: float = 0.5,
 ) -> Path:
@@ -1440,8 +939,6 @@ def mix_song(
         The audio format of the mixed song.
     output_name : str, optional
         The name of the mixed song.
-    display_msg : str, default="[~] Mixing audio tracks..."
-        The message to display when mixing the audio tracks.
     progress_bar : gr.Progress, optional
         Gradio progress bar to update.
     percentage : float, default=0.5
@@ -1452,66 +949,19 @@ def mix_song(
     Path
         The path to the song cover.
 
-    Raises
-    ------
-    NotProvidedError
-        If no audio tracks are provided.
-
     """
-    if not audio_track_gain_pairs:
-        raise NotProvidedError(
-            entity=Entity.AUDIO_TRACK_GAIN_PAIRS,
-            ui_msg=UIMessage.NO_AUDIO_TRACK,
-        )
-
-    audio_path_gain_pairs = [
-        (
-            to_wav(
-                _validate_exists(audio_track, Entity.AUDIO_TRACK),
-                song_dir,
-                "50_Input",
-                progress_bar=progress_bar,
-                percentage=percentage,
-            ),
-            gain,
-        )
-        for audio_track, gain in audio_track_gain_pairs
-    ]
-    song_dir_path = _validate_exists(song_dir, Entity.SONG_DIR)
-    args_dict = MixedSongMetaData(
-        staged_audio_tracks=[
-            StagedAudioMetaData(
-                audio_track=FileMetaData(
-                    name=audio_path.name,
-                    hash_id=get_file_hash(audio_path),
-                ),
-                gain=gain,
-            )
-            for audio_path, gain in audio_path_gain_pairs
-        ],
-        output_sr=output_sr,
-        output_format=output_format,
-    ).model_dump()
-
-    paths = [
-        get_unique_base_path(
-            song_dir_path,
-            "51_Mix",
-            args_dict,
-        ).with_suffix(suffix)
-        for suffix in ["." + output_format, ".json"]
-    ]
-
-    mix_path, mix_json_path = paths
-
-    if not all(path.exists() for path in paths):
-        display_progress(display_msg, percentage, progress_bar)
-
-        _mix_song(audio_path_gain_pairs, mix_path, output_sr, output_format)
-        json_dump(args_dict, mix_json_path)
+    mix_path = mix_audio(
+        audio_track_gain_pairs,
+        song_dir,
+        output_sr,
+        output_format,
+        content_type=MixedAudioType.SONG,
+        progress_bar=progress_bar,
+        percentage=percentage,
+    )
     output_name = output_name or get_song_cover_name(
-        audio_path_gain_pairs[0][0],
-        song_dir_path,
+        audio_track_gain_pairs[0][0],
+        song_dir,
         None,
     )
     song_path = OUTPUT_AUDIO_DIR / f"{output_name}.{output_format}"
@@ -1633,7 +1083,7 @@ def run_pipeline(
         intermediate audio files that were generated.
 
     """
-    _validate_exists(model_name, Entity.MODEL_NAME)
+    validate_exists(model_name, Entity.MODEL_NAME)
     display_progress("[~] Starting song cover generation pipeline...", 0, progress_bar)
     song, song_dir = retrieve_song(
         source,
@@ -1669,8 +1119,8 @@ def run_pipeline(
         percentage=3 / 9,
     )
     converted_vocals_track = convert(
-        vocals_track=vocals_dereverb_track,
-        song_dir=song_dir,
+        audio_track=vocals_dereverb_track,
+        directory=song_dir,
         model_name=model_name,
         n_octaves=n_octaves,
         n_semitones=n_semitones,
@@ -1680,14 +1130,15 @@ def run_pipeline(
         rms_mix_rate=rms_mix_rate,
         protect_rate=protect_rate,
         hop_length=hop_length,
-        split_vocals=split_vocals,
-        autotune_vocals=autotune_vocals,
+        split_audio=split_vocals,
+        autotune_audio=autotune_vocals,
         autotune_strength=autotune_strength,
-        clean_vocals=clean_vocals,
+        clean_audio=clean_vocals,
         clean_strength=clean_strength,
         embedder_model=embedder_model,
         embedder_model_custom=embedder_model_custom,
         sid=sid,
+        content_type=RVCContentType.VOCALS,
         progress_bar=progress_bar,
         percentage=4 / 9,
     )
@@ -1729,7 +1180,6 @@ def run_pipeline(
         output_sr,
         output_format,
         output_name,
-        display_msg="[~] Mixing main vocals, instrumentals, and backup vocals...",
         progress_bar=progress_bar,
         percentage=8 / 9,
     )
